Changes needed by the new framework, in progress

+++ dqm/analysis/modules/include/DQMHistAnalysisSVDEfficiency.h
@@ -65,9 +65,9 @@ namespace Belle2 {
   private:
 
     //parameters
-    float m_effError; /**<error level of the efficiency */
-    float m_effWarning; /**< warning level of the efficiency */
-    float m_statThreshold; /**<minimal number of tracks per sensor to set green or red frame */
+    double m_effError; /**<error level of the efficiency */
+    double m_effWarning; /**< warning level of the efficiency */
+    double m_statThreshold; /**<minimal number of tracks per sensor to set green or red frame */
     //! Data members
 
     /** Reference Histogram Root file name */
